Add files via upload
## Dataset

The dataset (`TSLA.csv`) included in this repository contains historical stock market data for Tesla Inc. This dataset is utilized in the stock price prediction project and includes information such as Open, High, Low, and Close prices, and trading volume. The data spans a specific timeframe and serves as the foundation for training and evaluating the machine learning models within the project. The dataset was obtained from Kaggle and is preprocessed for use in the provided Google Colab notebooks and source code.